# backend/contracts/api/openapi.yaml
openapi: 3.1.0
info:
  title: Comfort Curators V0 API Contract Floor
  version: 0.1.0
  description: >-
    Minimum public and operational REST surface for the Comfort Curators V0
    modular monolith. Domain requirements and authorization rules in the
    protected product files remain normative.
servers:
  - url: http://localhost:8080
security:
  - bearerAuth: []
...
paths:
...
  /v1/sessions/{session_id}/revocations:
    post:
      operationId: revokeSession
      tags: [Identity]
      parameters: [{$ref: '#/components/parameters/SessionId'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '202': {$ref: '#/components/responses/Accepted'}
        '404': {$ref: '#/components/responses/Error'}
  /v1/tenants/{tenant_id}/support-access-grants:
    post:
      operationId: createSupportAccessGrant
      tags: [Tenancy]
      parameters: [{$ref: '#/components/parameters/TenantId'}]
      requestBody: {$ref: '#/components/requestBodies/Command'}
      responses:
        '201': {$ref: '#/components/responses/Resource'}
        '403': {$ref: '#/components/responses/Error'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  parameters:
    TenantId: {name: tenant_id, in: path, required: true, schema: {$ref: '#/components/schemas/Identifier'}}
...
    SessionId: {name: session_id, in: path, required: true, schema: {$ref: '#/components/schemas/Identifier'}}
...
  requestBodies:
    Command:
      required: true
      content:
        application/json:
          schema: {$ref: '#/components/schemas/CommandEnvelope'}
...
  responses:
...
    Accepted:
      description: Durable command accepted
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Accepted'}
    Resource:
      description: Resource representation
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Resource'}
...
    Error:
      description: Stable error response
      content:
        application/json:
          schema: {$ref: '#/components/schemas/Error'}
  schemas:
    Identifier:
      type: string
      minLength: 16
      maxLength: 64
      pattern: '^[A-Za-z0-9_-]+$'
    CommandEnvelope:
      type: object
      additionalProperties: false
      required: [idempotency_key, data]
      properties:
        idempotency_key: {type: string, minLength: 8, maxLength: 128}
        reason: {type: string, maxLength: 1000}
        data: {type: object, additionalProperties: true}
...
    Accepted:
      type: object
      additionalProperties: false
      required: [request_id, status, resource_id]
      properties:
        request_id: {$ref: '#/components/schemas/Identifier'}
        status: {type: string, enum: [accepted, queued]}
        resource_id: {$ref: '#/components/schemas/Identifier'}
    Resource:
      type: object
      required: [id, version, data]
      properties:
        id: {$ref: '#/components/schemas/Identifier'}
        version: {type: integer, minimum: 1}
        data: {type: object, additionalProperties: true}
...
    Error:
      type: object
      additionalProperties: false
      required: [request_id, code, message]
      properties:
        request_id: {$ref: '#/components/schemas/Identifier'}
        code: {type: string, pattern: '^[A-Z0-9_]+$'}
        message: {type: string, maxLength: 1000}
        details: {type: object, additionalProperties: true}